requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
mysqlclient==2.1.1
phonenumbers==8.12.52
Pillow==9.2.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
stripe==3.5.0
tzdata==2022.1
urllib3==1.26.10
